Ugh, got distracted
Tried to reorganize my files and rebuild quickly.  didn't work out.

Had to branch and move changes cause I got distracted attempting to implement this.

Probably shouldn't branch off a feature like this, but if I can get this sorted i'll merg then continue working on audio pass through.
