+++ Source/Core/Parameters.h
@@ -1,0 +1,9 @@
+#pragma once
+#include "../JuceIncludes.h"
+
+namespace Params {
+    // IDs as compile-time constants
+    inline const juce::Identifier GAIN_ID{ "gain" };
+    inline const juce::Identifier FILTER_FREQ_ID{ "cutoff" };
+
+ }
