--- Source/PluginProcessor.cpp
@@ -1,182 +1,0 @@
-/*
-  ==============================================================================
-
-    This file contains the basic framework code for a JUCE plugin processor.
-
-  ==============================================================================
-*/
-
-#include "PluginProcessor.h"
-#include "PluginEditor.h"
-
-//==============================================================================
-ToneCascadeAudioProcessor::ToneCascadeAudioProcessor()
-#ifndef JucePlugin_PreferredChannelConfigurations
-     : AudioProcessor (BusesProperties()
-                     #if ! JucePlugin_IsMidiEffect
-                      #if ! JucePlugin_IsSynth
-                       .withInput  ("Input",  juce::AudioChannelSet::stereo(), true)
-                      #endif
-                       .withOutput ("Output", juce::AudioChannelSet::stereo(), true)
-                     #endif
-                       )
-#endif
-{
-}
-
-ToneCascadeAudioProcessor::~ToneCascadeAudioProcessor()
-{
-}
-
-//==============================================================================
-const juce::String ToneCascadeAudioProcessor::getName() const
-{
-    return JucePlugin_Name;
-}
-
-bool ToneCascadeAudioProcessor::acceptsMidi() const
-{
-   #if JucePlugin_WantsMidiInput
-    return true;
-   #else
-    return false;
-   #endif
-}
-
-bool ToneCascadeAudioProcessor::producesMidi() const
-{
-   #if JucePlugin_ProducesMidiOutput
-    return true;
-   #else
-    return false;
-   #endif
-}
-
-bool ToneCascadeAudioProcessor::isMidiEffect() const
-{
-   #if JucePlugin_IsMidiEffect
-    return true;
-   #else
-    return false;
-   #endif
-}
-
-double ToneCascadeAudioProcessor::getTailLengthSeconds() const
-{
-    return 0.0;
-}
-
-int ToneCascadeAudioProcessor::getNumPrograms()
-{
-    return 1;   // NB: some hosts don't cope very well if you tell them there are 0 programs,
-                // so this should be at least 1, even if you're not really implementing programs.
-}
-
-int ToneCascadeAudioProcessor::getCurrentProgram()
-{
-    return 0;
-}
-
-void ToneCascadeAudioProcessor::setCurrentProgram (int index)
-{
-}
-
-const juce::String ToneCascadeAudioProcessor::getProgramName (int index)
-{
-    return {};
-}
-
-void ToneCascadeAudioProcessor::changeProgramName (int index, const juce::String& newName)
-{
-}
-
-//==============================================================================
-void ToneCascadeAudioProcessor::prepareToPlay (double sampleRate, int samplesPerBlock)
-{
-    lastSampleRate = sampleRate; 
-    lastBlockSize = samplesPerBlock;
-
-    // Use this method as the place to do any pre-playback
-    // initialisation that you need..
-}
-
-void ToneCascadeAudioProcessor::releaseResources()
-{
-    // When playback stops, you can use this as an opportunity to free up any
-    // spare memory, etc.
-}
-
-#ifndef JucePlugin_PreferredChannelConfigurations
-bool ToneCascadeAudioProcessor::isBusesLayoutSupported (const BusesLayout& layouts) const
-{
-  #if JucePlugin_IsMidiEffect
-    juce::ignoreUnused (layouts);
-    return true;
-  #else
-    // This is the place where you check if the layout is supported.
-    // In this template code we only support mono or stereo.
-    // Some plugin hosts, such as certain GarageBand versions, will only
-    // load plugins that support stereo bus layouts.
-    if (layouts.getMainOutputChannelSet() != juce::AudioChannelSet::mono()
-     && layouts.getMainOutputChannelSet() != juce::AudioChannelSet::stereo())
-        return false;
-
-    // This checks if the input layout matches the output layout
-   #if ! JucePlugin_IsSynth
-    if (layouts.getMainOutputChannelSet() != layouts.getMainInputChannelSet())
-        return false;
-   #endif
-
-    return true;
-  #endif
-}
-#endif
-
-void ToneCascadeAudioProcessor::processBlock (juce::AudioBuffer<float>& buffer, juce::MidiBuffer& midiMessages)
-{
-    juce::ScopedNoDenormals noDenormals;
-    const int numChannels = buffer.getNumChannels();
-    const int numSamples = buffer.getNumSamples();
-    const int numInputs = getTotalNumInputChannels();
-
-    for (int ch = numInputs; ch < numChannels; ++ch) {
-        buffer.clear(ch, 0, numSamples);
-    }
-
-    for (int ch = 0; ch < numInputs; ++ch) {
-        auto* data = buffer.getWritePointer(ch);
-    }
-    
-}
-
-//==============================================================================
-bool ToneCascadeAudioProcessor::hasEditor() const
-{
-    return true; // (change this to false if you choose to not supply an editor)
-}
-
-juce::AudioProcessorEditor* ToneCascadeAudioProcessor::createEditor()
-{
-    return new ToneCascadeAudioProcessorEditor (*this);
-}
-
-//==============================================================================
-void ToneCascadeAudioProcessor::getStateInformation (juce::MemoryBlock& destData)
-{
-    // You should use this method to store your parameters in the memory block.
-    // You could do that either as raw data, or use the XML or ValueTree classes
-    // as intermediaries to make it easy to save and load complex data.
-}
-
-void ToneCascadeAudioProcessor::setStateInformation (const void* data, int sizeInBytes)
-{
-    // You should use this method to restore your parameters from this memory block,
-    // whose contents will have been created by the getStateInformation() call.
-}
-
-//==============================================================================
-// This creates new instances of the plugin..
-juce::AudioProcessor* JUCE_CALLTYPE createPluginFilter()
-{
-    return new ToneCascadeAudioProcessor();
-}
